pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "font-audit-crawler"
version = "0.1.0"
description = "Deterministic Playwright-based crawler for runtime font compliance audits and font remediation QA."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "Mykhailo Yarytskyi" }]
keywords = ["fonts", "playwright", "crawler", "qa", "audit", "typography"]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: Apache Software License",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Testing",
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.4",
  "PyYAML>=6.0.1",
  "httpx>=0.27.0",
  "selectolax>=0.3.21",
  "Jinja2>=3.1.4",
  "orjson>=3.10.3",
  "playwright>=1.53.0",
  "python-slugify>=8.0.4",
]

[project.optional-dependencies]
dev = [
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.5.27",
  "mypy>=1.10.0",
  "pre-commit>=3.7.1",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "pytest-xdist>=3.6.1",
  "ruff>=0.5.0",
  "types-PyYAML>=6.0.12.20241230",
]

[project.scripts]
font-audit = "font_audit_crawler.cli:app"

[project.urls]
Homepage = "https://github.com/mmmihaeel/font-audit-crawler"
Documentation = "https://github.com/mmmihaeel/font-audit-crawler/tree/main/docs"
Issues = "https://github.com/mmmihaeel/font-audit-crawler/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
font_audit_crawler = ["rules/*.yaml", "reporting/templates/*.j2"]

[tool.pytest.ini_options]
addopts = "-ra --strict-markers --maxfail=1 --cov=src/font_audit_crawler --cov-report=term-missing"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
mypy_path = "src"
packages = ["font_audit_crawler"]

[tool.coverage.run]
source = ["src/font_audit_crawler"]

[tool.coverage.report]
show_missing = true
skip_covered = false
